backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlmodel==0.0.14
psycopg2-binary==2.9.9
pydantic==2.5.0
python-dotenv==1.0.0
pyjwt==2.8.0
pytest==7.4.3
httpx==0.25.2
pytest-asyncio==0.21.1
alembic==1.13.1
